Includes sample code from https://github.com/geal/nom-trace
In order to make the sample code work, I needed to use "nom_trace" instead
of "nom-trace" in the external crate line.  I also needed to use "nom::digit"
instead of "digit" in the source.

I probably should have committed before making those changes, but oh well.

FWIW, the output I get when I run it is different from the output in the
nom-trace README.md.  It looks to me like its reporting byte slices instead
of string slices.  I'm guessing this is because we start out with a
parser<Vec<&[u8]>>.  I'm going to see if I can get it to work with string
slices.