Interval parsing code is now in interval_parse.rs, lots of unused complaints.
I was going to turn a bunch of this into a library to not get the unused
complaints.  I'm not sure if I really want to do that or if I simply want to
start using the code I've written.

Turning it into a library might give me better insight into how modules
can be exposed by the root src/lib.rs.