Correctly detects intervals in all the files I've tested against.
I've tested it against all the intervals I've run starting on June 25th,
which is when I reduced my rest period to 30 seconds.  Before that I was
running intervals from fixed places and resting a lot more in between.
So, rather than implement a fairly robust detector, I make use of the fact
that intervals are close to one another.  That works well enough to give
me something I can actually use on a week to week basis, but doesn't allow
me to retroactively look back before June 25th (not that I want to!).

This code is still pretty hack-and-slash, and although it outputs the
average of all the intervals it finds, it doesn't report some of the
other stats that I typically look at on Mondays.  OTOH, adding that
functionality shouldn't be hard.