Don't panic if no intervals are found.
I had put together some text files that had unexpected decimal points in
them and that caused the intervals to not be recognized.  That led to a
divide by zero panic.

This commit prevents that divide by zero, but doesn't address the issue of
the decimal places (which I think should have been allowed; I'll double
check now).