[Feat] computes meters-per-sec where it's not in the GPX file.  Closes #12.
For calculating meters-per-second I simply take the haversine_length
and divide by duration. This does not take elevation into
consideration, but for my use of this software, I don't think that
matters enough to attempt using elevation (which is, I believe,
significantly less accurate than latitude and longitude).

If meters-per-sec is already present in all the Trkpts, then we
analyze twice, once with the supplied meters-per-sec and then again
with the computed meters-per-sec.

The difference between the average interval pace between supplied and
computed meters-per-second for any given day can be considerable,
computed meters-per-second for which may mean that there's a bug
somewhere.  The numbers that are generated by this software aren't
sufficiently important to me to merit further investigation any time
soon.